requirements.txt
autopep8==1.7.0
build==0.8.0
click==8.1.3
dnspython==2.2.1
flake8==5.0.4
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
pep517==0.13.0
pycodestyle==2.9.1
pyflakes==2.5.0
pymongo==4.2.0
pyparsing==3.0.9
python-dotenv==0.20.0
toml==0.10.2
tomli==2.0.1
Werkzeug==2.2.2
